requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.14.0
gunicorn==23.0.0
packaging==24.2
pytz==2024.2
setuptools==75.1.0
sqlparse==0.5.3
tzdata==2024.2
wheel==0.44.0
